requirements.txt
beautifulsoup4
requests
selenium
google-generativeai
webdriver-manager